Use dedicated copy_wait/copy_image_wait functions, implement wrappers in __init__.py
